moving method to appropriate place
